+++ src/main/java/org/scoula/consumption/service/ConsumptionService.java
@@ -6,13 +6,13 @@ import org.scoula.consumption.dto.ExpectedReqDTO;
 
 public interface ConsumptionService {
 
-    ConsumptionCalendarDTO getCal(Long memberNo, String yearMonth);
+    ConsumptionCalendarDTO getCal(Integer memberNo, String yearMonth);
 
-    void addExpected(Long memberNo, ExpectedReqDTO request);
+    void addExpected(Integer memberNo, ExpectedReqDTO request);
 
     void updateExpected(Long expectedNo, ExpectedReqDTO request);
 
     void deleteExpected(Long expectedNo);
 
-    void detectRecurringSpending(Long memberNo);
+    void detectRecurringSpending(Integer memberNo);
 }
